Refactor import paths and remove unused files
- Updated import paths for supabase client and utility functions to reflect new directory structure.
- Removed obsolete supabase-client.js and utils.js files.
- Cleaned up validation logic by moving it to a new location.
- Ensured consistent usage of imports across components.
