+++ App/frontend/src/components/ui/form.jsx
@@ -3,7 +3,7 @@ import * as React from "react"
 import { Slot } from "@radix-ui/react-slot"
 import { Controller, FormProvider, useFormContext } from "react-hook-form";
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/utils/utils"
 import { Label } from "@/components/ui/label"
 
 const Form = FormProvider
