+++ App/frontend/src/components/ui/input.jsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/utils/utils"
 
 const Input = React.forwardRef(({ className, type, ...props }, ref) => {
   return (
