+++ App/frontend/src/components/ui/label.jsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import * as LabelPrimitive from "@radix-ui/react-label"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/utils/utils"
 
 function Label({
   className,
