+++ App/frontend/src/hooks/authForm.jsx
@@ -1,12 +1,11 @@
-import {useState} from "react";
-import {supabase} from "../supabase-client";
-import {validPassword, validEmail} from "../lib/validation";
+import { useState } from "react";
+import { supabase } from "../lib/supabaseClient";
+import { validEmail, validPassword } from "../utils/validation";
 
 export const AuthForm = (mode) => {
     const [email, setEmail] = useState("");
     const [password, setPassword] = useState("");
     const [message, setMessage] = useState("");
-
 
     const validate = () => {
         if (!email || !password) {
@@ -34,16 +33,16 @@ export const AuthForm = (mode) => {
 
         let result;
         if (mode === "register") {
-            result = await supabase.auth.signUp({email, password});
+            result = await supabase.auth.signUp({ email, password });
             setMessage(result.error ? result.error.message : "Registration successful! A confirmation link has been sent to your email.");
 
 
         } else if (mode === "login") {
-            result = await supabase.auth.signInWithPassword({email, password});
+            result = await supabase.auth.signInWithPassword({ email, password });
             setMessage(result.error ? "Wrong email or password." : "");
         }
     };
 
-    return {email, setEmail, password, setPassword, message, handleSubmit};
+    return { email, setEmail, password, setPassword, message, handleSubmit };
 };
