+++ App/frontend/src/pages/notesPage.jsx
@@ -1,4 +1,4 @@
-import { supabase } from "../supabase-client";
+import { supabase } from "../lib/supabaseClient";
 import { Button } from "@/components/ui/button";
 
 const NotesPage = () => {
