--- App/frontend/src/supabase-client.js
@@ -1,9 +1,0 @@
-import { createClient } from '@supabase/supabase-js'
-
-const supabaseUrl = import.meta.env.VITE_SUPABASE_URL
-const supabaseKey = import.meta.env.VITE_SUPABASE_ANON_KEY
-
-export const supabase = createClient(
-  supabaseUrl,
-  supabaseKey
-);
